refactor: extract window features
